requirements.txt
openai
pyfiglet
termcolor
python-dotenv